requirement.txt
stremlit
fpdf
python-docx
Pillow